requirements.txt
streamlit
google-generativeai
python-dotenv
langchain
PyPDF2
PyMuPDF
pdf2image
chromadb
faiss-cpu
langchain_google_genai
langchain_community
docx2txt
anthropic
sentence-transformers
pytesseract 
pillow
easyocr 